fix UScaledNs.MarshalBinary() wrong cast

+++ ptp.go
@@ -147,7 +147,7 @@ func NewUScaledNs(b []byte) (UScaledNs, error) {
 func (p *UScaledNs) MarshalBinary() ([]byte, error) {
 	b := make([]byte, UScaledNsLen)
 
-	binary.BigEndian.PutUint16(b[:4], uint16(p.ms))
+	binary.BigEndian.PutUint32(b[:4], uint32(p.ms))
 
 	binary.BigEndian.PutUint64(b[4:], p.ls)
